Added VLAN field, modified remotePort and VNC, removed instances serialization
