+++ src/Entity/Lab.php
@@ -43,12 +43,6 @@ class Lab implements InstanciableInterface
     private $devices;
 
     /**
-     * @ORM\OneToMany(targetEntity="App\Entity\LabInstance", mappedBy="lab", cascade={"persist", "remove"})
-     * @Serializer\Groups({"lab", "start_lab", "stop_lab"})
-     */
-    private $instances;
-
-    /**
      * @ORM\ManyToOne(targetEntity="App\Entity\User", inversedBy="createdLabs")
      * @Serializer\Groups({"lab", "lab_author"})
      */
@@ -83,9 +77,7 @@ class Lab implements InstanciableInterface
         $this->devices = new ArrayCollection();
         $this->connexions = new ArrayCollection();
         $this->activities = new ArrayCollection();
-        $this->instances = new ArrayCollection();
         $this->uuid = (string) new Uuid();
-        $this->deviceInstances = new ArrayCollection();
         $this->networkInterfaceInstances = new ArrayCollection();
         $this->createdAt = new \DateTime();
         $this->lastUpdated = new \DateTime();
@@ -146,64 +138,6 @@ class Lab implements InstanciableInterface
     {
         if ($this->devices->contains($device)) {
             $this->devices->removeElement($device);
-        }
-
-        return $this;
-    }
-
-    /**
-     * @return Collection|Instance[]
-     */
-    public function getInstances()
-    {
-        return $this->instances;
-    }
-
-    /**
-     * Returns the instance corresponding to user.
-     *
-     * @param User $user
-     * @return LabInstance|null
-     */
-    public function getUserInstance(User $user): ?LabInstance
-    {
-        $criteria = Criteria::create()->where(Criteria::expr()->eq("user", $user));
-
-        $instance = $this->instances->matching($criteria)->first();
-
-        return $instance ?: null;
-    }
-
-    public function addInstance(Instance $instance): self
-    {
-        if (!$this->instances->contains($instance)) {
-            $this->instances[] = $instance;
-        }
-
-        return $this;
-    }
-
-    public function removeInstance(Instance $instance): self
-    {
-        if ($this->instances->contains($instance)) {
-            $this->instances->removeElement($instance);
-        }
-
-        return $this;
-    }
-
-    public function hasDeviceUserInstance(User $user): bool
-    {
-        return $this->devices->exists(function ($index, Device $device) use ($user) {
-            return $device->getUserInstance($user) != null;
-        });
-    }
-
-    public function setInstances(array $instances): self
-    {
-        $this->getInstances()->clear();
-        foreach ($instances as $instance) {
-            $this->addInstance($instance);
         }
 
         return $this;
